requirements.txt
flask>=2.0.0
requests>=2.27.1
python-dotenv>=0.20.0
pydantic>=1.9.0
flask-cors>=3.0.10